+++ ChangeLog.md
@@ -2,6 +2,12 @@
 
 All notable changes are documented in this file using the [Keep a CHANGELOG](https://keepachangelog.com/) principles.
 
+## [3.2.0] - 2023-MM-DD
+
+### Added
+
+* `ComplexityCollection::sortByDescendingCyclomaticComplexity()`
+
 ## [3.1.0] - 2023-09-28
 
 ### Added
@@ -48,6 +54,7 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 
 * Initial release
 
+[3.2.0]: https://github.com/sebastianbergmann/complexity/compare/3.1.0...main
 [3.1.0]: https://github.com/sebastianbergmann/complexity/compare/3.0.1...3.1.0
 [3.0.1]: https://github.com/sebastianbergmann/complexity/compare/3.0.0...3.0.1
 [3.0.0]: https://github.com/sebastianbergmann/complexity/compare/2.0.2...3.0.0